+++ src/main/java/lotto/validator/PurchaseAmountValidator.java
@@ -1,0 +1,13 @@
+package lotto.validator;
+
+public class PurchaseAmountValidator {
+    public static void validatePurchaseAmount(final String input) {
+        if (InputValidator.isEmpty(input)) {
+            throw new IllegalArgumentException(InputValidator.ENTER_VALUE_MESSAGE);
+        }
+
+        if (!InputValidator.isMatchedNumberRegex(input)) {
+            throw new IllegalArgumentException(InputValidator.INVALID_INPUT_CHARACTER);
+        }
+    }
+}
