+++ src/main/java/lotto/view/InputView.java
@@ -5,7 +5,9 @@ import static camp.nextstep.edu.missionutils.Console.readLine;
 import java.util.Arrays;
 import java.util.List;
 import lotto.domain.dto.PurchaseAmountDto;
-import lotto.validator.InputValidator;
+import lotto.validator.BonusNumberValidator;
+import lotto.validator.PurchaseAmountValidator;
+import lotto.validator.WinningLottoValidator;
 
 public class InputView {
     private static final String INPUT_PURCHASE_AMOUNT_MESSAGE = "구입금액을 입력해 주세요.";
@@ -15,7 +17,7 @@ public class InputView {
     public PurchaseAmountDto inputPurchaseAmount() {
         System.out.println(INPUT_PURCHASE_AMOUNT_MESSAGE);
         String input = readLine();
-        InputValidator.validatePurchaseAmount(input);
+        PurchaseAmountValidator.validatePurchaseAmount(input);
 
         return new PurchaseAmountDto(Integer.parseInt(input));
     }
@@ -23,7 +25,7 @@ public class InputView {
     public List<Integer> inputWinningNumbers() {
         System.out.println(INPUT_WINNING_NUMBERS_MESSAGE);
         String input = readLine();
-        InputValidator.validateWinningNumbers(input);
+        WinningLottoValidator.validateWinningNumbers(input);
 
         return Arrays.stream(input.split(", ")).map(Integer::parseInt).toList();
     }
@@ -31,7 +33,7 @@ public class InputView {
     public int inputBonusNumber() {
         System.out.println(INPUT_BONUS_NUMBER_MESSAGE);
         String input = readLine();
-        InputValidator.validateBonusNumber(input);
+        BonusNumberValidator.validateBonusNumber(input);
 
         return Integer.parseInt(input);
     }
